requirements.txt
-i https://pypi.org/simple
click==8.1.3; python_version >= '3.7'
flask==2.1.2
importlib-metadata==4.11.3; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
mysql-connector-python==8.0.29
protobuf==3.20.1; python_version >= '3.7'
werkzeug==2.1.2; python_version >= '3.7'
zipp==3.8.0; python_version >= '3.7'
